String together and debug dashboard services
Everything is working on a basic level - no interaction with the
server or db, and only using a simple-ass mock event db. Which works!

I think the order I'm going to go in is setting up the components
for the event list and resource viewer before trying to wire up
everything all over again, at least the skeletal version of the components.
In two days I should have a rough sketch of the dashboard which I'll
just need to add funtionality to. I'm liking the feel of this a lot more
than angular 1.5, incidentally. Or maybe I'm just enjoying the process more
because I feel more comfortable with it! Who can say. But I think things are
going well!