Scaffold event-viewer, begin adding datepicker
My first real attempt to make ng2 do something! I should have known
it would be a little more complicated than I had thought. Right now
I want to initialize the datepicker in a div set aside for that
purpose, but it's a little more complex than that because apparently
in ng2 you aren't supposed to touch the DOM with a 10 foot pole. So
Im about to find out if the dom-querying api in ng2 is rich enough to
let me start this datepicker I found, haha
than that because apparently in ng2