Event-list-item wired up
- event-list-item now ngFors itself away in event-list scroll.
Missing the more complicated behavior with the days and the sorting
and all that but hey it's early days.

- Patched a handful of style issues with event-list-item - in
particular, text-overflow hadn't been working properly.

- I used a very, uh, 'simplistic' way to pass the event buffer to
the event-list-scroll, by which I mean I literally handed it down
through my ridiculous ladder of components until it got where it
needed to be. I should try a service - or the event service itself,
once I can figure out how to make sure it gets initialized before
I try to pull events out of it. That sounds like a job for promises,
or for what observables appear to be, but whatever right now it's
kludgy but it works.