WIP - still working on Observables and EventService
- Rough plan I've settled on: BackendService.getEvents returns
Observable<Event>, which is subscribed to in EventService.
Subscription sorts each event into an EventBuffer (Event[][]), which
is the data structure used to represent events split up by days.
Its state is subject to change and it is managed by EventService.

( Open question: Does the EventService eventBuffer begin at one date
and grow in size indefinitely, or does it have a flexible first date
and remain fixed? The second one sounds like a better representation
for the Dashboard component to have of the EventService's data, so maybe
the EventService should have the first representation. Also,
does the EventBuffer data structure allow the effective use of local
storage? )

EventService's eventBuffer's state can change as a result of two things:
- adding more events to the eventBuffer
- *updating the events in eventBuffer ( can be solved later )

EventService will have an addEvents toBuffer Method which uses same
backend.getEvents public method, but - actually the same subscription
might work here, just extend the EventService eventBuffer farther from origin.
( -- oh but here's a corner case: past events. eventBuffer with
origin at today's date is no good for that. But also, past
events shouldn't be able to be updated or edited -
maybe it's for the best if they're represented in a different
EventBuffer? But now you've got more complicated state management
to do. Hm)

EventService will also have an updateEvents(rangeStart, rangeEnd,
timeOfLastRequest) method, which makes a request to the backend
for only the events which have changed since the last request,
in an Observable. (Is this feasible to implement on the server?
How would it know when events were last changed?)