WIP - adding events to displayed scheduler
- the scheduler itself is styled fine as it is, so woohoo. And
I've lost all the jQuery/fc kruft so that's cool, but now I'm
trying to figure out the best way to display these events.
Inserting a Component or creating a DOM node are both apparently
frowned upon, but I'm not sure what the best approach here is
for doing things the Angular Way. Probably just need to do more
planning before diving in like this again.