Calendar works (!!!)
- It was really hard not to type that commit header in all caps,
so I'm going to do that here:
THE CALENDAR WORKS HA HA HA WOW

- It breaks when you change days, altho I thought I had ironed out
a bug which made it do that. It could be a lingering effect of the
fact that the calendar is initialized on today.

- Changed data structure of timeSlotList/timeSlot ( consider type
aliasing typeSlot? ) to numbers representing minutes from midnight
rather than Moments. Should be more performant (?) but most importantly
avoids the issue of trying to update the moment dates in the timeSlotList
for each day so that the minute comparisons aren't off by multiples of 24 * 60.