Nested observable data structure working?
- Although it's possible every call to getEventsByDay() traverses
the whole damn Observable< GroupedObservable< EmEvent > > anyway
despite using Observable.first as an operator, which you would think
doesn't do that. But it's hard to tell how exactly RxJs works under the
hood, if my past few mins of console logging made anything clear.

- Uh where was I aside from complaining oh yes, event.service now
correctly sorts events into their sub-Observables by day, and exposes
a public method to access a day's events as Observable<EmEvent>.
That's what I wanted! And if it turns out to be a bad idea, not a
waste of time because I think I'm just barely maybe starting to grok
Observables. We'll find out when it comes to wiring up the rest of the
app!