~Browser Inconsistency~ with web animations API polyfill
- While doing browser testing, realized that web animations API chugs
horribly on everything except chrome. It looks like to get the same
effect we'll need to work a little harder, but I don't know how to
create manual transition fallbacks in the case of angular's animation
not working, so for now just disabled the plugin. No animations looks
better than the janky animations right now.