Working build/min process
- wow, that was really, really hard. Now using rollup, which
works based on es6 modules, so that took a lot of fiddling. Also,
porting the things I needed from client side node_modules was a little
(read: very) confusing, and the solution currently (just using a gulp
script to pick them up and stick them in dist) is a little brittle.